refactor: extract managed terminal account
